+++ core/libs/editor/src/index.ts
@@ -21,7 +21,9 @@ export * from './lib/editor.service';
 export * from './lib/editor.types';
 export * from './lib/helpers/form-field.helper';
 export * from './lib/large-file-warning-dialog/large-file-warning-dialog.service';
+export * from './lib/model-checker.service';
 export * from './lib/model-loader.service';
+export * from './lib/model-saver.service';
 export * from './lib/open-element-window/open-element-window.component';
 export * from './lib/open-element-window/open-element-window.service';
 export * from './lib/rename-model/rename-model.component';
