+++ core/libs/editor/src/lib/model-loader.service.ts
@@ -14,15 +14,13 @@
 import {ModelApiService} from '@ame/api';
 import {LoadedFilesService, NamespaceFile} from '@ame/cache';
 import {InstantiatorService} from '@ame/instantiator';
-import {ExporterHelper} from '@ame/migrator';
 import {RdfModelUtil} from '@ame/rdf/utils';
-import {BrowserService, ElectronSignalsService, ModelSavingTrackerService, NotificationsService, config} from '@ame/shared';
-import {FileStatus} from '@ame/sidebar';
+import {BrowserService, ElectronSignalsService, ModelSavingTrackerService, NotificationsService} from '@ame/shared';
 import {Injectable, inject} from '@angular/core';
-import {ModelElementCache, NamedElement, RdfModel, Samm, loadAspectModel} from '@esmf/aspect-model-loader';
+import {ModelElementCache, NamedElement, RdfModel, loadAspectModel} from '@esmf/aspect-model-loader';
 import {RdfLoader} from 'libs/aspect-model-loader/src/lib/shared/rdf-loader';
 import {NamedNode} from 'n3';
-import {Observable, Subject, catchError, forkJoin, map, of, switchMap, tap, throwError} from 'rxjs';
+import {Observable, catchError, forkJoin, map, of, switchMap, tap, throwError} from 'rxjs';
 import {ModelRendererService} from './model-renderer.service';
 import {LoadModelPayload} from './models/load-model-payload.interface';
 import {LoadingCodeErrors} from './models/loading-errors';
@@ -156,62 +154,6 @@ export class ModelLoaderService {
           return acc;
         }, {}),
       ),
-    );
-  }
-
-  /**
-   * Gets all files from workspace and process if they have any error or missing dependencies
-   *
-   * @param signal used to get the current file in process
-   * @returns
-   */
-  detectWorkspaceErrors(signal?: Subject<string>): Observable<Record<string, FileStatus>> {
-    let namespacesStructure: Record<string, string[]>;
-
-    const extractDependencies = (absoluteName: string) =>
-      this.modelApiService.getAspectMetaModel(absoluteName).pipe(
-        switchMap(rdf => this.parseRdfModel([rdf])),
-        map(rdfModel => {
-          const dependencies = RdfModelUtil.resolveExternalNamespaces(rdfModel)
-            .map(external => external.replace(/urn:samm:|urn:bamm:|#/gi, ''))
-            .filter(dependency => !`urn:samm:${dependency}`.includes(Samm.BASE_URI));
-
-          const missingDependencies: string[] = [];
-
-          for (const dependency of dependencies) {
-            const [namespace, version] = dependency.split(':');
-            if (!namespacesStructure[`${namespace}:${version}`]) missingDependencies.push(dependency);
-          }
-          const [, , fileName] = absoluteName.split(':');
-          const status = new FileStatus(fileName);
-          const currentFile = this.loadedFilesService.currentLoadedFile;
-
-          status.dependencies = dependencies;
-          status.missingDependencies = missingDependencies;
-          status.unknownSammVersion = rdfModel.samm.version === 'unknown';
-          status.outdated = ExporterHelper.isVersionOutdated(rdfModel.samm.version, config.currentSammVersion);
-          status.loaded = currentFile?.absoluteName === absoluteName;
-          status.errored = status.unknownSammVersion || Boolean(status.missingDependencies.length);
-
-          signal?.next(absoluteName);
-
-          return status;
-        }),
-      );
-
-    return this.modelApiService.getNamespacesStructure().pipe(
-      switchMap(structure => {
-        namespacesStructure = structure;
-        const requests = {};
-        for (const namespace in structure) {
-          for (const file of structure[namespace]) {
-            const absoluteName = `${namespace}:${file}`;
-            requests[absoluteName] = extractDependencies(absoluteName);
-          }
-        }
-
-        return forkJoin(requests);
-      }),
     );
   }
